Windows build with VS2017

+++ src/ledger/LedgerManagerImpl.h
@@ -6,9 +6,9 @@
 #include "util/asio.h"
 
 #include "history/HistoryManager.h"
-#include "ledger/SyncingLedgerChain.h"
 #include "ledger/LedgerHeaderFrame.h"
 #include "ledger/LedgerManager.h"
+#include "ledger/SyncingLedgerChain.h"
 #include "main/PersistentState.h"
 #include "transactions/TransactionFrame.h"
 #include "util/Timer.h"
@@ -79,13 +79,14 @@ class LedgerManagerImpl : public LedgerManager
     uint32_t getLedgerNum() const override;
     uint32_t getLastClosedLedgerNum() const override;
     int64_t getMinBalance(uint32_t ownerCount) const override;
-    int64_t getTxFee() const override;
+    uint32_t getTxFee() const override;
     uint32_t getMaxTxSetSize() const override;
     uint64_t getCloseTime() const override;
     uint64_t secondsSinceLastLedgerClose() const override;
     void syncMetrics() override;
 
-    void startNewLedger(int64_t balance, uint32_t baseFee, uint32_t baseReserve, uint32_t maxTxSetSize);
+    void startNewLedger(int64_t balance, uint32_t baseFee, uint32_t baseReserve,
+                        uint32_t maxTxSetSize);
     void startNewLedger() override;
     void loadLastKnownLedger(
         std::function<void(asio::error_code const& ec)> handler) override;
